修复超大权重size超过int上限的问题

+++ include/devices/tfacc/fastllm-tfacc.h
@@ -27,7 +27,7 @@ namespace fastllm {
 
         void Wait();
 
-        void SendLongMessage(uint8_t *buffer, int len);
+        void SendLongMessage(uint8_t *buffer, uint64_t len);
 
         void RegisterFastllmData(fastllm::Data *data, const std::string &weightType);
